feat: constants 파일 통합 및 import 수정 (#54)
