+++ constants/game.ts
@@ -10,4 +10,20 @@ export const GAME_GUIDE_COMMENT = '최적의 타이밍에 거래해서 우승을
 
 export const DAY_OPTIONS = ['10일', '15일', '20일']
 
+export const INITIAL_COINS = 1000
 export const INITIAL_FISH_PRICE = 100
+export const INITIAL_TIMER = 20
+
+// 더미 데이터
+export const TOTAL_ROUNDS = 3
+export const FINAL_COIN = 100
+
+export const AVATARS = [
+  { imageUrl: '/images/cat-1.png', nickName: '대장고양이', isLeader: true, id: '1' },
+  { imageUrl: '/images/cat-2.png', nickName: '제임스', id: '2' },
+  { imageUrl: '/images/cat-3.png', nickName: '레드히어로', id: '3' },
+  { imageUrl: '/images/cat-4.png', nickName: '마크정식주세요제발요', id: '4' },
+  { imageUrl: '/images/cat-5.png', nickName: '다은메롱', id: '5' },
+]
+
+export const SIX_AVATARS = [...AVATARS, ...Array(Math.max(0, 6 - AVATARS.length)).fill({})]
